+++ lib/runners/processor/code_sniffer.rb
@@ -5,23 +5,19 @@ module Runners
     Schema = _ = StrongJSON.new do
       # @type self: SchemaClass
 
+      let :target, enum?(string, array(string))
+
       let :runner_config, Runners::Schema::BaseConfig.base.update_fields { |fields|
         fields.merge!({
-                        version: enum?(string, numeric),
-                        dir: string?,
-                        standard: string?,
-                        extensions: enum?(string, array(string)),
-                        encoding: string?,
-                        ignore: enum?(string, array(string)),
-                        # DO NOT ADD ANY OPTION under `options`.
-                        options: optional(object(
-                                            dir: string?,
-                                            standard: string?,
-                                            extensions: string?,
-                                            encoding: string?,
-                                            ignore: string?
-                                          ))
-                      })
+          version: enum?(string, numeric),
+          target: target,
+          dir: target, # alias for `target`
+          standard: enum?(string, array(string)),
+          extensions: enum?(string, array(string)),
+          encoding: string?,
+          ignore: enum?(string, array(string)),
+          parallel: boolean?,
+        })
       }
 
       let :issue, object(
@@ -33,30 +29,26 @@ module Runners
 
     register_config_schema(name: :code_sniffer, schema: Schema.runner_config)
 
-    DefaultOptions = _ = Struct.new(:standard, :extensions, :dir, keyword_init: true)
-    OPTIONS_CAKE_PHP = DefaultOptions.new(standard: "CakePHP", extensions: "php", dir: "app/").freeze
-    OPTIONS_SYMFONY = DefaultOptions.new(standard: "Symfony", extensions: "php", dir: "src/").freeze
-    OPTIONS_PSR2 = DefaultOptions.new(standard: "PSR2", extensions: "php", dir: "./").freeze
+    DefaultOptions = _ = Struct.new(:standard, :extensions, :target, keyword_init: true)
+    OPTIONS_CAKE_PHP = DefaultOptions.new(standard: "CakePHP", extensions: "php", target: "app/").freeze
+    OPTIONS_SYMFONY = DefaultOptions.new(standard: "Symfony", extensions: "php", target: "src/").freeze
+    OPTIONS_PSR2 = DefaultOptions.new(standard: "PSR2", extensions: "php", target: "./").freeze
 
     def self.config_example
       <<~'YAML'
         root_dir: project/
-        dir: app/
-        standard: CakePHP
+        target: [app/, test/]
+        standard: [CakePHP, custom-ruleset.xml]
         extensions: [php, inc]
         encoding: utf-8
         ignore:
           - app/vendor/
+        parallel: true
       YAML
     end
 
     def analyzer_bin
       "phpcs"
-    end
-
-    def setup
-      add_warning_if_deprecated_options
-      yield
     end
 
     def analyze(changes)
@@ -68,7 +60,7 @@ module Runners
         "--runtime-set", "ignore_errors_on_exit", "1", # See https://github.com/squizlabs/PHP_CodeSniffer/wiki/Configuration-Options#ignoring-errors-when-generating-the-exit-code
         "--runtime-set", "ignore_warnings_on_exit", "1", # See https://github.com/squizlabs/PHP_CodeSniffer/wiki/Configuration-Options#ignoring-warnings-when-generating-the-exit-code
         *additional_options,
-        directory,
+        *Array(config_linter[:target] || config_linter[:dir] || default_options.target),
       )
 
       issues = []
@@ -101,36 +93,19 @@ module Runners
 
     def additional_options
       if config_linter.empty?
-        ["--standard=#{default_options.standard}", "--extensions=#{default_options.extensions}"]
+        [
+          "--standard=#{default_options.standard}",
+          "--extensions=#{default_options.extensions}",
+        ]
       else
-        [standard_option, extensions_option, *encoding_option, *ignore_option]
+        [
+          "--standard=#{comma_separated_list(config_linter[:standard] || default_options.standard)}",
+          "--extensions=#{comma_separated_list(config_linter[:extensions] || default_options.extensions)}",
+          *(config_linter[:encoding].then { |enc| enc ? ["--encoding=#{enc}"] : [] }),
+          *(comma_separated_list(config_linter[:ignore]).then { |list| list ? ["--ignore=#{list}"] : [] }),
+          *(config_linter[:parallel] ? ["--parallel=2"] : []),
+        ]
       end
-    end
-
-    def standard_option
-      standard = config_linter[:standard] || config_linter.dig(:options, :standard) || default_options.standard
-      "--standard=#{standard}"
-    end
-
-    def extensions_option
-      extensions = config_linter[:extensions] || config_linter.dig(:options, :extensions) || default_options.extensions
-      extensions = comma_separated_list(extensions)
-      "--extensions=#{extensions}"
-    end
-
-    def encoding_option
-      encoding = config_linter[:encoding] || config_linter.dig(:options, :encoding)
-      encoding ? ["--encoding=#{encoding}"] : []
-    end
-
-    def ignore_option
-      ignore = config_linter[:ignore] || config_linter.dig(:options, :ignore)
-      ignore = comma_separated_list(ignore)
-      ignore ? ["--ignore=#{ignore}"] : []
-    end
-
-    def directory
-      config_linter[:dir] || config_linter.dig(:options, :dir) || default_options.dir
     end
 
     def default_options