+++ test/smokes/code_sniffer/expectations.rb
@@ -37,15 +37,20 @@ s.add_test(
       git_blame_info: {
         commit: :_, line_hash: "48ee9fdb6490aadc28bbb90cee75350be54532de", original_line: 6, final_line: 6
       }
+    },
+    {
+      path: "test/test.php",
+      location: { start_line: 6, start_column: 1 },
+      id: "PSR2.Files.ClosingTag.NotAllowed",
+      message: "A closing tag is not permitted at the end of a PHP file",
+      links: [],
+      object: { type: "ERROR", severity: 5, fixable: true },
+      git_blame_info: {
+        commit: :_, line_hash: "48ee9fdb6490aadc28bbb90cee75350be54532de", original_line: 6, final_line: 6
+      }
     }
   ],
-  analyzer: { name: "PHP_CodeSniffer", version: default_version },
-  warnings: [
-    {
-      message: /The `linter.code_sniffer.options` option is deprecated/,
-      file: "sideci.yml"
-    }
-  ]
+  analyzer: { name: "PHP_CodeSniffer", version: default_version }
 )
 
 s.add_test("with_php_version", type: "success", issues: :_, analyzer: { name: "PHP_CodeSniffer", version: default_version })
@@ -142,6 +147,36 @@ s.add_test(
   issues: [
     {
       path: "src/foo.php",
+      location: { start_line: 2, start_column: 1 },
+      id: "PSR2.Files.ClosingTag.NotAllowed",
+      message: "A closing tag is not permitted at the end of a PHP file",
+      links: [],
+      object: { type: "ERROR", severity: 5, fixable: true },
+      git_blame_info: {
+        commit: :_, line_hash: "48ee9fdb6490aadc28bbb90cee75350be54532de", original_line: 2, final_line: 2
+      }
+    }
+  ],
+  analyzer: { name: "PHP_CodeSniffer", version: default_version }
+)
+
+s.add_test(
+  "option_parallel",
+  type: "success",
+  issues: [
+    {
+      path: "a.php",
+      location: { start_line: 2, start_column: 1 },
+      id: "PSR2.Files.ClosingTag.NotAllowed",
+      message: "A closing tag is not permitted at the end of a PHP file",
+      links: [],
+      object: { type: "ERROR", severity: 5, fixable: true },
+      git_blame_info: {
+        commit: :_, line_hash: "48ee9fdb6490aadc28bbb90cee75350be54532de", original_line: 2, final_line: 2
+      }
+    },
+    {
+      path: "b.php",
       location: { start_line: 2, start_column: 1 },
       id: "PSR2.Files.ClosingTag.NotAllowed",
       message: "A closing tag is not permitted at the end of a PHP file",